requirements.txt
# Documentation Download Agent Requirements
# Python 3.8+ recommended

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Async HTTP client
aiohttp>=3.9.1

# HTML parsing
beautifulsoup4>=4.12.2
lxml>=4.9.3

# File operations
aiofiles>=23.2.1
PyYAML>=6.0.1

# Type hints
typing-extensions>=4.8.0

# AI / Vector Search
sentence-transformers>=2.2.2
numpy>=1.24.0

# MCP (Model Context Protocol)
mcp

# Development dependencies (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0